ed_vise/requirements.txt
google-api-python-client==X.X.X
pip >= 23.2
django >= 5.0.3
google-auth-oauthlib
google-api-python-client